+++ cybsi/api/user/enums.py
@@ -25,15 +25,15 @@ class RoleName(CybsiAPIEnum):
     FeedAdministrator = "FeedAdministrator"
     """
         FeedAdministrator's role permissions:
-        [DataSources:r, Feeds:rw, FeedsData:r, Observable:r,
-        ReputationLists:rw, ReputationListsContent:r, Search:r,
-        SearchFilters:rw, StoredQuery:rw, Users:r]
+        [DataSources:r, Observable:r, ReputationLists:rw,
+        ReputationListsContent:r, Search:r, StoredQuery:rw,
+        Users:r]
     """
     FeedDataReader = "FeedDataReader"
     """
         FeedDataReader's role permissions:
-        [DataSources:r, Feeds:r, FeedsData:r, ReputationLists:r,
-        ReputationListsContent:r, SearchFilters:r, StoredQuery:r, Users:r]
+        [DataSources:r, ReputationLists:r, ReputationListsContent:r,
+        StoredQuery:r, Users:r]
     """
     EnrichmentRunner = "EnrichmentRunner"
     """
@@ -83,7 +83,7 @@ class RoleName(CybsiAPIEnum):
     Searcher = "Searcher"
     """
         Searcher's role permissions:
-        [DataSources:r, Observable:r, Search:r, SearchFilters:rw]
+        [DataSources:r, Observable:r, Search:r]
     """
     UserAdministrator = "UserAdministrator"
     """